Fix stale-barrier event handling bug.
On a ScreenChange event we remove and forget about all the barriers we
set up, but we might still have hit/leave events for those barriers
queued up waiting to be handled, leading to crashes when we tried to
look up those barriers and failed to find them.

We now just treat lookup failure as an expected condition that can arise
sometimes for legitimate reasons (this one) instead of treating it as an
unconditional internal error.